升级包版本并修改类型，删除无用属性
升级 `Graphics.Core.csproj` 中的包版本从 `2.1.7` 到 `2.1.9`。
在 `ImGuiController.cs` 和 `ImGuiRenderer.cs` 中，将 `GetBinding` 和 `RemoveBinding` 方法的返回类型和参数类型从 `nint` 改为 `ulong`。
在 `ImGuiRenderer.cs` 中，将 `_mapped`、`_mappedTextures`、`_selfSets` 和 `_selfViews` 字典的键类型从 `nint` 改为 `ulong`。
在 `Joystick.cs` 中，删除了 `Direction` 属性和相关代码。

+++ Source/Graphics/Graphics.Vulkan/ImGui/ImGuiController.cs
@@ -174,17 +174,17 @@ public unsafe class ImGuiController : DisposableObject
         }
     }
 
-    public nint GetBinding(ResourceFactory factory, TextureView textureView)
+    public ulong GetBinding(ResourceFactory factory, TextureView textureView)
     {
         return _imGuiRenderer.GetBinding(factory, textureView);
     }
 
-    public nint GetBinding(ResourceFactory factory, Texture texture)
+    public ulong GetBinding(ResourceFactory factory, Texture texture)
     {
         return _imGuiRenderer.GetBinding(factory, texture);
     }
 
-    public void RemoveBinding(nint binding)
+    public void RemoveBinding(ulong binding)
     {
         _imGuiRenderer.RemoveBinding(binding);
     }
